Move the jsdom installation logic out of the sbt plugin
